Add multiprocessing.
Files are now searched in parallel.  prefind.prefind.finder is
refactored to use concurrent.futures.ProcessPoolExecutor when searching
through files, since a single file search is "embarrassingly parallel".

At the same time this makes searching a few small files significantly
slower (in test cases from ~.012 s to ~.055 s).  So, in case of small
searches, it should still be quite reasonable, but faster for lots of
big files.  The slowdown in unit tests, since their running time order
of magnitude is so small, could also be due to using sets instead of
lists for file paths.  We use these to make sure there are no duplicate
paths to search, and that their order is ignored.

Because concurrent.futures isn't in Python 2(.7), Prefind is no longer
compatible with Python 2.  Reflect that in README.